Move agent and workflow managers into services
Agent and workflow manager ownership now lives under the service tree.

Canonical agent manager API:

    import "github.com/valon-technologies/gestalt/server/services/agents/agentmanager"

    manager := agentmanager.New(agentmanager.Config{...})
    turn, err := manager.CreateTurn(ctx, principal, coreagent.ManagerCreateTurnRequest{...})

Canonical workflow manager API:

    import "github.com/valon-technologies/gestalt/server/services/workflows/workflowmanager"

    manager := workflowmanager.New(workflowmanager.Config{...})
    run, err := manager.StartRun(ctx, principal, workflowmanager.RunStart{...})

The package names stay agentmanager and workflowmanager; this commit moves the directories and rewrites imports without changing manager behavior.
