+++ gestaltd/internal/bootstrap/agent_workflow_tools_test.go
@@ -16,8 +16,8 @@ import (
 	"github.com/valon-technologies/gestalt/server/internal/agentgrant"
 	"github.com/valon-technologies/gestalt/server/internal/principal"
 	"github.com/valon-technologies/gestalt/server/internal/registry"
-	"github.com/valon-technologies/gestalt/server/internal/workflowmanager"
 	"github.com/valon-technologies/gestalt/server/services/invocation"
+	"github.com/valon-technologies/gestalt/server/services/workflows/workflowmanager"
 )
 
 func TestAgentRuntimeWorkflowSystemToolCreatesScopedSchedule(t *testing.T) {
