+++ gestaltd/internal/bootstrap/lazy_workflow_manager.go
@@ -7,7 +7,7 @@ import (
 
 	coreworkflow "github.com/valon-technologies/gestalt/server/core/workflow"
 	"github.com/valon-technologies/gestalt/server/internal/principal"
-	"github.com/valon-technologies/gestalt/server/internal/workflowmanager"
+	"github.com/valon-technologies/gestalt/server/services/workflows/workflowmanager"
 )
 
 type lazyWorkflowManager struct {
