+++ gestaltd/services/agents/agentmanager/manager.go
