+++ gestaltd/services/agents/agentmanager/manager_test.go
