+++ gestaltd/services/agents/agentmanager/tool_search.go
