+++ gestaltd/services/agents/agentmanager/turn_display.go
